Added proper sign in checks

+++ app/(kambaz)/courses/[cid]/modules/page.tsx
@@ -8,31 +8,33 @@ import ModulesControls from "./ModulesControls";
 import ModuleControlButtons from "./ModuleControlButtons";
 import LessonControlButtons from "./LessonControlButtons";
 import { useState, useEffect } from "react";
-import {
-  setModules,
-  addModule,
-  editModule,
-  updateModule,
-  deleteModule,
-} from "./reducer";
+import { setModules, editModule, updateModule } from "./reducer";
 import { useSelector, useDispatch } from "react-redux";
 import * as client from "../../client";
 import { RootState } from "../../../store";
 export default function Modules() {
   const { cid } = useParams();
   const dispatch = useDispatch();
+  const { currentUser } = useSelector(
+    (state: RootState) => state.accountReducer,
+  ) as any;
+  const canEditModules = ["FACULTY", "TA"].includes(currentUser?.role);
+
   const onCreateModuleForCourse = async () => {
+    if (!canEditModules) return;
     if (!cid) return;
     const newModule = { name: moduleName, course: cid };
     const module = await client.createModuleForCourse(cid as string, newModule);
     dispatch(setModules([...modules, module]));
   };
   const onRemoveModule = async (moduleId: string) => {
+    if (!canEditModules) return;
     await client.deleteModule(moduleId);
     dispatch(setModules(modules.filter((m: any) => m._id !== moduleId)));
   };
 
   const onUpdateModule = async (module: any) => {
+    if (!canEditModules) return;
     await client.updateModule(module);
     const newModules = modules.map((m: any) =>
       m._id === module._id ? module : m,
@@ -49,10 +51,6 @@ export default function Modules() {
   }, [cid, dispatch]);
   const [moduleName, setModuleName] = useState("");
   const { modules } = useSelector((state: RootState) => state.modulesReducer);
-  const { currentUser } = useSelector(
-    (state: RootState) => state.accountReducer,
-  ) as any;
-  const canEditModules = currentUser && currentUser.role !== "STUDENT";
   return (
     <div>
       <>
